+++ apps/web/src/pages/Departments/ViewDepartmentPage.tsx
@@ -10,7 +10,7 @@ import {
   Link,
   Card,
   CardSeparator,
-  List,
+  Ul,
 } from "@gc-digital-talent/ui";
 import {
   FragmentType,
@@ -104,7 +104,7 @@ export const ViewDepartmentForm = ({ query }: ViewDepartmentProps) => {
             {department.orgIdentifier ?? notProvided}
           </FieldDisplay>
           <FieldDisplay label={intl.formatMessage(labels.departmentType)}>
-            <List.UL unStyle space="sm">
+            <Ul unStyle space="sm">
               <li>
                 <BoolCheckIcon value={department.isCorePublicAdministration}>
                   {intl.formatMessage(labels.corePublicAdmin)}
@@ -125,7 +125,7 @@ export const ViewDepartmentForm = ({ query }: ViewDepartmentProps) => {
                   {intl.formatMessage(labels.regulatory)}
                 </BoolCheckIcon>
               </li>
-            </List.UL>
+            </Ul>
           </FieldDisplay>
           <FieldDisplay label={intl.formatMessage(labels.departmentSize)}>
             {department.size?.label.localized ?? notProvided}
